Make Instance a top-level class
This is a code health/style nit pick. Since Instance dataclass does
not *need* to be class attribute of the RegistrySingleton, I don't think
that it should be. Code maintanence and testing will be easier in the
future if Instance grows to be a more complex class. Additionally, if a
new class is created that also wants to use Instance (say a new type of
registry), then you wouldn't want Instance so tightly coupled to a
single registry implementation.