Hash objects rather than names to avoid collisions
The RegistrySingleton._registy dictionary used the __name__ of objects
as the key. This is potentially fragile because classes or functions
with the same name but defined in different modules would use the same
key. In Python, classes and functions are hashable anyways so we can
just use those as the key directly.

An alternative would be to use the full path, rather than just the name.

def fullpath(t: type[Any]) -> str:
    return f'{t.__module__}.{t.__qualname__}'