Fix unreliable object ID comparisions in tests
In Python 3.7, the assert id(obj) != obj_id statements in the tests can
fail. This happens because the original obj used to get obj_id = id(obj)
is overwritten. Once this happens, the original obj can be garbage
collected, and since id() uses the memory address of objects, the
previously allocated memory of the original object can be allocated to
one of the new objects that gets assigned to the obj variable. Then when
we compare the id of the new object assigned to obj, it matches the id
of the original object because they are both allocated the same memory
block.

I've implemented the simplest solution here, which is just to hold a
reference for the scope of the test to the original obj used to create
the obj_id value. This means no latter objects can be assigned the same
memory address and therefore have the same id.